api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
python-dotenv==1.0.1
boto3==1.37.25
qdrant-client==1.12.0
numpy==1.26.4